src/requirements.txt
requests==2.28.2
python-dotenv
pymongo
pandas